move schema into form page
